Addressed issue with mobile on 'work-content' grid
I created a 'reverse-columns' class to flip the grid layout. I want the images and text to alternate on each row when on desktop. But on mobile, I just want them to appear stacked. Finally addressed the issue. I laid out all my html in the same order and I managed all the ordering of the elements through CSS